Introduce `pkg/utils/gardener/operator` package
