+++ pkg/utils/gardener/operator/extension.go
@@ -1,8 +1,8 @@
-// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
+// SPDX-FileCopyrightText: 2025 SAP SE or an SAP affiliate company and Gardener contributors
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package gardener
+package operator
 
 import (
 	"fmt"
